requirements.txt
# requirements file used by mybinder.org to execute the examples
dash
flake8
frozendict
fsspec
ipykernel
keras
matplotlib
numpy
pandas
pytest
pyyaml
requests
seaborn
setuptools
scikit-learn
tensorflow 
torch
torchvision
coverage
jupyter
# azure file system
adlfs
pyarrow
s3fs
